Added cart Service and Controller Still some methods remail to implement

+++ src/main/java/com/greenscan/dto/response/CartResponse.java
@@ -22,8 +22,8 @@ public class CartResponse {
     private UserResponse vendor;
     private UserResponse pickupAssistant;
     
-    private Long lattiude;
-    private Long logitude;
+    private Double lattiude;
+    private Double logitude;
     private String pickupAddress;
     
     private String pickupInstructions;
